refactor(nodelist): remove banned/bannedBy lists
This removes the `banned` and `bannedBy` collections that were used to
track which nodes were banned or have banned us. Instead, we use the
`banned` and `bannedBy` attributes of each node directly. This is done
to simplify the code and prevent possible inconsistencies between the
collections and the attributes on the node.