feat(config): property type and value checking
This adds assertion logic to the `Config` class to validate that any
properties specified are of the expected type and, for certain
properties, that the value is valid. This may prevent crashes or other
unexpected behavior in xud due to invalid configuration, and will also
log more instructive errors when a specified property is invalid.

Closes #1222.